requirements.txt
flask==3.1.0
Werkzeug==3.1.3
flask-socketio==5.5.1
gevent==25.4.2
python-dotenv==1.1.0
openai[datalib,voice_helpers]==1.78.0
numpy==2.2.5
pydub==0.25.1
ffmpeg-python==0.2.0
httpx==0.28.1
requests==2.32.3
gevent-websocket==0.10.1
python-socketio==5.13.0
RPi.GPIO==0.7.1
sounddevice==0.5.1
pytest==8.3.5
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-cov==6.1.1google-genai
